refactor(db): move getInternalID into lishs-schema to break barrel import cycle
